chore: move pattern object to constants and make hint an object
