--- src/validations/index.ts
@@ -1,3 +1,0 @@
-import * as Validations from "./validations";
-
-export { Validations };
